Moved code to get mqtt credentials to config instead of dispatcher
